refactor(mcp-memory): split rag.ts into focused modules (#26)
Split the 3149-line monolithic rag.ts into 12 focused modules under
src/rag/ while preserving the public API surface unchanged.

Modules: types, utils, config, repository, chunking, bm25, semantic,
embedding, weaviate, reranker, providers, knowledge-base.
