+++ packages/mcp-memory/src/rag/index.ts
@@ -1,0 +1,24 @@
+export type {
+  KnowledgeBaseSource,
+  OpenVikingConfig,
+  KnowledgeBaseConfig,
+  EmbeddingConfig,
+  RerankerConfig,
+  VectorStoreConfig,
+  WeaviateVectorStoreConfig,
+  RagMetadataFilter,
+  RagQueryParams,
+  RagQueryMatch,
+  RagQueryTiming,
+  RagQueryResult,
+  KnowledgeProvider,
+} from './types.js';
+
+export {
+  createKnowledgeBase,
+  ragQuery,
+  normalizeKnowledgeBaseSource,
+  normalizeOpenVikingConfig,
+  normalizeOpenVikingMatches,
+  ragQuerySchema,
+} from './providers.js';
