+++ packages/shared/src/security/shell-analysis.test.ts
