+++ tsconfig.json
@@ -23,4 +23,3 @@
   },
   "exclude": ["node_modules", "dist"]
 }
-
